database for an RPG game described on ekfwnhsh.pdf
It is a database for an RPG game called Gok. (it is an imaginary one)

The concept is that we have some info, like
Each player can have many characters in the game. A game has a wide
range
of characters, a character can have one or more pets to accompany him.
He (the character) is a part of a tribe, has some abilities, and can
carry items, and gold coins.
The abilities can be natural (strength, speed etc) or magical (different
type of
spells). Items have a specific type, and can be either weapons, shields
or whatever
a character can aquire during the game.
The character has stats, different worlds exist etc etc.

Based on that information we designed the relational database.

Full description of this is in Greek, in the file ekfwnhsh.pdf